Create a "source-file" function in .rc/rc.
source-file() takes the path to a file to source and sources it, if it
exists.  It is simply a wrapper to make sourcing files prettier.  If
RC_VERBOSE is defined, it will also notify the user which file it is
sourcing, which means that the individual scripts do not have to do this.