tput syntax is wonky.
I don't honestly know the ins and outs of tput, but for whatever reason
on Linux (and macOS?) calling "tput 0 0 0" works, but OpenBSD takes
offence to that and just wants "tput 0". So the quick, "I have no idea
what's going on" solution is just to make two different if blocks for
each behavior.